Added scene loading from a text file and refactored some code to accomodate that new system

+++ include/matrix-math.h
@@ -2,11 +2,11 @@
 #define MATRIX_MATH_H
 #include "custom-types.h"
 
-Vec3 MatrixVec3Multiplication(Vec3 vec, int matrixSize, double matrix[matrixSize][matrixSize]);
+Vec3 MatrixVec3Multiplication(Vec3 vec, int matrixSize, float matrix[matrixSize][matrixSize]);
 
-Vec4 MatrixVec4Multiplication(Vec4 vec, int matrixSize, double matrix[matrixSize][matrixSize]);
+Vec4 MatrixVec4Multiplication(Vec4 vec, int matrixSize, float matrix[matrixSize][matrixSize]);
 
 Vec3 AddVec3(Vec3 vec1, Vec3 vec2);
 
-Vec3 RotateVec3AroundAxis(Vec3 vec, double angle, RotationAxis axis);
+Vec3 RotateVec3AroundAxis(Vec3 vec, float angle, RotationAxis axis);
 #endif
