+++ src/matrix-math.c
@@ -1,6 +1,6 @@
 #include "../include/matrix-math.h"
 
-Vec3 MatrixVec3Multiplication(Vec3 vec, int matrixSize, double  matrix[matrixSize][matrixSize]) {
+Vec3 MatrixVec3Multiplication(Vec3 vec, int matrixSize, float  matrix[matrixSize][matrixSize]) {
     Vec3 ret = { 0, 0, 0 };
 
     ret.x += (
@@ -24,7 +24,7 @@ Vec3 MatrixVec3Multiplication(Vec3 vec, int matrixSize, double  matrix[matrixSiz
     return ret;
 }
 
-Vec4 MatrixVec4Multiplication(Vec4 vec, int matrixSize, double matrix[matrixSize][matrixSize]) {
+Vec4 MatrixVec4Multiplication(Vec4 vec, int matrixSize, float matrix[matrixSize][matrixSize]) {
     Vec4 ret = { 0, 0, 0, 0 };
 
     ret.x += (
@@ -62,11 +62,11 @@ Vec3 AddVec3(Vec3 vec1, Vec3 vec2) {
     return (Vec3) {vec1.x + vec2.x, vec1.y + vec2.y, vec1.z + vec2.z};
 }
 
-Vec3 RotateVec3AroundAxis(Vec3 vec, double angle, RotationAxis axis) {
+Vec3 RotateVec3AroundAxis(Vec3 vec, float angle, RotationAxis axis) {
     switch (axis) {
         case X_AXIS:
             return MatrixVec3Multiplication(vec, 3, 
-                (double[3][3]){
+                (float[3][3]){
                     {1, 0, 0},
                     {0, cos(angle), - sin(angle)},
                     {0, sin(angle), cos(angle)},
@@ -74,7 +74,7 @@ Vec3 RotateVec3AroundAxis(Vec3 vec, double angle, RotationAxis axis) {
             );
         case Y_AXIS:
             return MatrixVec3Multiplication(vec, 3, 
-                (double[3][3]){
+                (float[3][3]){
                     {cos(angle), 0, sin(angle)},
                     {0, 1, 0},
                     {- sin(angle), 0, cos(angle)},
@@ -82,7 +82,7 @@ Vec3 RotateVec3AroundAxis(Vec3 vec, double angle, RotationAxis axis) {
             );
         case Z_AXIS:
             return MatrixVec3Multiplication(vec, 3, 
-                (double[3][3]){
+                (float[3][3]){
                     {cos(angle), - sin(angle), 0},
                     {sin(angle), cos(angle), 0},
                     {0, 0, 1 },
